all changes set

+++ src/main/java/com/search/teacher/service/FileService.java
@@ -1,18 +1,13 @@
 package com.search.teacher.service;
 
 import com.search.teacher.dto.ImageDto;
-import com.search.teacher.dto.modules.listening.ListeningResponse;
 import com.search.teacher.exception.BadRequestException;
 import com.search.teacher.exception.NotfoundException;
 import com.search.teacher.model.entities.Image;
-import com.search.teacher.model.entities.Organization;
 import com.search.teacher.model.entities.User;
-import com.search.teacher.model.entities.modules.reading.ReadingQuestion;
 import com.search.teacher.model.enums.ImageType;
 import com.search.teacher.model.response.JResponse;
 import com.search.teacher.repository.ImageRepository;
-import com.search.teacher.service.modules.ListeningService;
-import com.search.teacher.service.organization.OrganizationService;
 import com.search.teacher.service.user.UserService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
@@ -33,59 +28,31 @@ public class FileService {
     private final UserService userService;
     private final ImageRepository imageRepository;
     private final FileUploadService fileUploadService;
-    private final OrganizationService organizationService;
 
     public JResponse uploadPhoto(User currentUser, MultipartFile file, String type) {
         ImageDto image = fileUploadService.fileUpload(currentUser, file, type, true);
         return JResponse.success(image);
     }
 
-    public String checkFilename(String audio, ImageType imageType) {
-        Image image = Optional.ofNullable(imageRepository.findByObjectNameAndImageType(audio, imageType))
-            .orElseThrow(() -> new BadRequestException("Please upload Listening Audio"));
-
-        return image.getUrl();
-    }
-
-    public Image saveImageToQuestion(String imageUrl) {
-        return Optional.ofNullable(imageRepository.findByObjectNameAndImageType(imageUrl, ImageType.READING))
-            .orElseThrow(() -> new BadRequestException("Please upload Reading Passage"));
-    }
-
-    public JResponse getAudioUrl(User currentUser, Long id) {
-        Image image = Optional.ofNullable(imageRepository.findByIdAndUserId(id, currentUser.getId()))
-            .orElseThrow(() -> new NotfoundException("Audio file not found"));
-
-        ImageDto imageDto = new ImageDto(image.getId(), image.getUrl(), image.getOriginalFilename());
-        return JResponse.success(imageDto);
-    }
-
     public JResponse deleteAudioFile(User currentUser, Long id) {
         Image image = Optional.ofNullable(imageRepository.findByIdAndUserId(id, currentUser.getId()))
-            .orElseThrow(() -> new NotfoundException("Audio file not found"));
+                .orElseThrow(() -> new NotfoundException("Image file not found"));
 
         fileUploadService.removeObject(image, image.getBucket());
         return JResponse.success();
     }
 
-    public Image audioUpload(User currentUser, MultipartFile file) {
-        Image image = new Image();
-        image.setImageType(ImageType.AUDIO);
-        image.setSize(file.getSize());
-        image.setActive(true);
-        image.setUserId(currentUser.getId());
-        image.setOriginalFilename(file.getOriginalFilename());
-        imageRepository.save(image);
-        fileUploadService.uploadAudio(file, image);
-        imageRepository.save(image);
-        return image;
-    }
-
-    public void removeAudio(Long userId, Image audio) {
-        Image image = Optional.ofNullable(imageRepository.findByIdAndUserId(audio.getId(), userId))
-            .orElseThrow(() -> new NotfoundException("Audio file not found"));
-
-        fileUploadService.removeObject(image, image.getBucket());
-        imageRepository.delete(image);
+    public JResponse deleteAudioByUrl(User currentUser, String url) {
+        Image image = imageRepository.findByUrl(url);
+        if (image != null) {
+            if (image.getUserId().equals(currentUser.getId())) {
+                fileUploadService.removeObject(image, image.getBucket());
+                imageRepository.delete(image);
+                return JResponse.success();
+            } else {
+                throw new NotfoundException("Image file not found");
+            }
+        }
+        return JResponse.error(404, "Image file not found");
     }
 }
